Restore the original behaviour of ProjectGenerator tests
Summary:
Before the flag `GENERATE_TARGETS_FOR_DEPENDENCIES` was removed, tests in
`ProjectGeneratorTest` used it, which effectively meant that the generated
project included all of the targets in the specified graph and ignored the
`initialTargets` parameter. When the flag was removed,
the projects generated in the tests only included the targets explicitly
specified in `initialTargets`. To restore the original behaviour, the
`initialTargets` parameter is now removed and automatically set to include
targets for all nodes of the passed graph.

Test Plan: Run tests.