Use the StubJar class to generate ABI jars for java_library and its ilk.
Summary:
The original way of constructing an ABI for a jar was by compiling it with a custom
annotation processor which sucked out the signature. Rather than adding an annotation
processor, we now have the `StubJar` class, which can generate a stable jar
representing the same thing.

This makes minimal changes to the tests for the existing ABI infrastructure, but a follow
up diff will delete that tree and rename this one.

Test Plan: buck test --all