New utility class PathListing
Summary:
In D1679650, @coneko asked me to refactor all the existing
code that knows how to list, sort, and filter files.

This diff introduces a new class `PathListing` which knows how to:

1) List files in a directory
2) Sort them
3) Apply optional filters (size and/ot count)

The class can work in two modes, either including only files which
match the filters, or excluding files which match the filters.

I plan on replacing the logic from `ChromeTraceBuildListener`
and `TracesHelper` with this logic.

This class is in its own Buck rule to avoid dependency
messes.

Test Plan: Unit tests included.