PathListing: Don't throw exceptions if we can't get last modified time
Summary:
Previously, `PathListing` contained a `Comparator` which
called `Files.getLastModifiedTime()`. However, this method throws
`IOException` if the file doesn't exist, so we had to translate
that to `RuntimeException`.

This caused problems if files were deleted out from underneath
`PathListing` by an external process: it would throw
`RuntimeException`, which of course nobody would catch.

The behavior we want in `PathListing` in this case is simply
to drop the offending entries from the result, so this diff:

1) Changes `PathListing` from accepting a `Comparator<Path>` to
accepting a `Function<Path, FileTime>` (so we can override the
behavior from `ProjectFilesystem`)

2) Uses the passed-in function to build a map containing the last modified
times of all paths matching the glob pattern

3) Sorts and returns the resulting map by value

Test Plan: `ant java-test`