Add a utility class to locate debug sections in native file formats
Summary:
Debug sections frequently contain environment specific data that needs
to be stripped or sanitized to improve the cachability of native code.
This diff adds a utlity class to locate relevant debug sections from
a native file format.  For now, this is just supports ELF, but a macho
parser is on the way.

Test Plan: added unittest and various object file data