Don't delete watchman watches when cleaning up buckd
Summary:
I confirmed with @wez, the author of watchman, that
deleting Watchman watches is unnecessary. In memory, watches
are represented very compactly, and they are automatically
removed if the watched directory is deleted.

Deleting and recreating watches is extremely expensive.  When watchman
isn't running, we were actually launching watchman, rescanning the
entire source tree, then deleting and re-creating the watch -- a
process which could take upwards of 30 seconds on large repositories.

This simply removes the code which deletes Watchman watches when
cleaning up buckd.

Test Plan:
`buck targets; buckd --kill; watchman watch-list; buck targets` in a repo
with lots of files. Before, watch was deleted and this sequence took
a very long time. Now, watch is not deleted and sequence is quick.