unify AppleTestDescription into a single node instead of three.
Summary:
Instead of generating a dylib, bundle, and test node when creating a test,
generate only a test node that includes fields of the dylib and bundle.

When creating build rules, the bundle and dylib rules are generated dynamically
from the info present in the test target using "graph enhancement" in the same
manner as `android_binary` et al.

This reduces the need for the python code to generate multiple targets and test
code to replicate the python and generate multiple targets. This also removes
the need to distinguish between nodes that are turned into targets and nodes
that are not. Finally (the reason i'm doing this), it simplifies work needed to
allow test rules to be instantiated as either a dylib bundle or a static lib,
since the exact implementation is not fixed.

Test Plan:
See that it generates functionally identical pbxproj and xcscheme files
in existing codebase.