Correctly indicate that inner classes are inner classes when creating the ABI jar.
Summary:
Although the ABI was being calculated sufficiently well for us to be able
to detect changes in public ABI by hashing the jar, it wasn't possible to
compile code against the abi jars when calling inner classes. This change
adds a test to verify that this is possible.

Test Plan: buck test --all