Optionally set the bootclasspath from buckconfig
Summary:
When specifying a source level, it is recommended that the bootclasspath
is set. The value of the bootclasspath depends on the source level so it
should be clear that setting this globally isn't a great idea.

This diff allows a bootclasspath to be configured per java source level,
using a config section such as:

  [java]
  bootclasspath-7 = some/clever/path.jar:someother.jar

No validation of this bootclasspath is done.

Test Plan:
buck test --all

Take a large codebase with Java 6 dependencies and compile with Java 8. Before this change, the compilation with Java 8 fails. After it, the compilation succeeds.