[nailgun client] Cherry-pick isatty forwarding diff 6f66e4a from upstream
Summary:
Buck's detection of when to enable or disable ANSI escape sequences
is based on checking if `System.console()` returns `null`.

This is a bad check for a few reasons:

1) It is based on the state of the JVM when `buckd` was launched, not
the current terminal the `buck` client is connected to, and

2) It means we cannot launch `buckd` with stdio file descriptors closed,
which means `ssh hostname buck build blah` hangs forever waiting for
the `buckd` process to close its stdio file descriptors

To fix this, we need the nailgun client to tell the server whether
its stdio file descripors are connected to TTYs. Luckily, this upstream
diff already implemented exactly that:

https://github.com/martylamb/nailgun/pull/20

This diff cherry-picks that change into our local copy of ng.c.

Test Plan:
buck build buck, added logs to confirm new environment
variables were passed