[buckd] Augment ANSI System.console() != null check with Nailgun TTY check
Summary:
Buck's detection of when to enable or disable ANSI escape sequences
is based on checking if `System.console()` returns `null`.

This is a bad check for a few reasons:

1) It is based on the state of the JVM when `buckd` was launched, not
the current terminal the `buck` client is connected to, and

2) It means we cannot launch `buckd` with stdio file descriptors closed,
which means `ssh hostname buck build blah` hangs forever waiting for
the `buckd` process to close its stdio file descriptors

This builds on previous commit to read the `isatty()` checks performed by
the nailgun client in `buckd` so each client separately enables or
disables ANSI escape sequences as appropriate.

Test Plan:
Set `com.facebook.buck.util.level = FINER` in `.bucklogging.local.properties`.
Ran the following tests and confirmed in logs the expected values present
and that ANSI escape sequences were or were not emitted as appropriate:

  NO_BUCKD=1 buck build buck            # with ANSI
  NO_BUCKD=1 TERM=dumb buck build buck  # no ANSI
  NO_BUCKD=1 buck build buck 2>/tmp/out # no ANSI
  TERM=dumb buck build buck             # no ANSI, buckd launched
  buck build buck                       # with ANSI, same buckd
  buck build buck 2>/tmp/out            # no ANSI, same buckd