Add and use a `Tool` abstraction to represent C/C++ tools
Summary:
Currently, we hash the contents of the various C/C++ tools we use
(e.g. cpp, g++, gcc, ar, ld, etc) into the rules that use them
(e.g. `CxxLink`, `CxxPreprocess`).  This is a sane default behavior
but isn't sufficient in some cases.  The pressing case is if we have
two tools which differ in their contents but which we *know* generate
identical output.

For example, the Android NDK has different compilers for each build
platform, but which generate the same code.  If we use the contents of
the compiler binary, we miss the opportunity to share the cache
between builds running on linux and Mac OS.  In this case, we'd like
to only include the NDK version (and perhaps some other interesting
settings) into the rule key.

Another case is if the binary contents of the tool alone aren't enough
to determine how it will run.  For example, a tool may read settings
from a configuration file.  In this case, we should also add (a subset
of) the configuration to the rule key.

This diff adds a `Tool` class to abstract these tools to address this,
which allows the tool to decide how it contributes to the rule keys,
currently supporting two modes:
1) The original, default, behavior is to hash in the contents of tool.
2) Alternately, the `Tool` instance can be constructed with a string to
use in liu of the contents when calculating the rule key.

This diff also changes the `NdkCxxPlatform` implementation to use option
(2), and generates this version string using the NDK version and the
target platform.

Test Plan:
- Added a unittest for the `Tool` class.
- Built NDK libs on Mac OS and Linux and verified they had the same rule key.