Easy: New enum ApplePlatform
Summary:
Unlike for Android, when building for Apple, we need to choose
between Mac OS X, iPhone OS, or iPhone Simulator, which have separate SDKs.

We also need to switch on the platform to decide whether to add
`-miphoneos-version-min=...` or `-mmacosx-version-min=...` parameters
to CFLAGS in `AppleCxxPlatform`.

This adds a new enum `ApplePlatform` which lets us indicate the platform
the user is building for.

Test Plan: Used in a subsequent diff.