Easy: New immutable value type AppleSdk
Summary:
Currently, `AppleSdkDiscovery` returns `Map<String, AppleSdkPaths>`
where the `String` is the name of the Apple SDK that was discovered.

We need more information about the SDK to add the proper compilation
flags to CFLAGS:

1) The SDK version
2) The platform (Mac OS X, iPhone OS, or iPhone Simulator)
3) The architecture(s) supported by the SDK

This diff adds a new value type, `AppleSdk`, which includes the name
plus these three pieces of information.

Currently, `CompilationDatabase` extracts 1) from the name of the SDK,
but this isn't guaranteed to work. A subsequent diff will replace its
logic with calling a getter on this object.

Test Plan: Used in subsequent diff.