Change TargetsCommand to use the target graph.
Summary:
TargetsCommand currently works with the action graph. This causes
it to output information about synthesised nodes, even though the
user is generally not aware of them as they're not addressable.
This changes it to use the target graph instead. The options to
show the rule key and output path of rules only work when
specifying targets, and still use the action graph.

Test Plan: CI, new tests