Create AppleCxxPlatforms in KnownBuildRuleTypes
Summary:
Now that we have `AppleCxxPlatform`, we can create it in `KnownBuildRuleTypes` so we can feed it down to the various Apple `Description` implementations.

This diff allows `KnownBuildRuleTypes` to optionally query the installed Apple SDK to see what platforms and SDKs are available, creates `AppleCxxPlatform`s for each, and adds them to the set of `CxxPlatform`s.

We don't yet pass them down to the `Description`s for Apple build rules; that's next.

Currently, we always assume the minimum SDK version we are targeting is the same as the version of the SDK. To fix this, instead of generating every possible `CxxPlatform` ahead of time, I plan on:

1) Adding a utility to parse a `Flavor` into key-value pairs
2) Creating a class which verifies the key-value pairs make sense for a target platform (say, Apple would want `sdk`, `arch`, and `version-min`—but Android might be different)
3) Creating a class which dynamically produces an `AppleCxxPlatform` based on the verified key-value pairs

Test Plan: buck test --all