Change AppleConfig.getAppleSdkPaths() to take ProcessExecutor instead of Console
Summary:
Uri suggested this cleanup when reviewing my change "Create
AppleCxxPlatforms in KnownBuildRuleTypes".

Doing this allows us to greatly simplify that change. Since we already
have access to `ProcessExecutor` in `KnownBuildRuleTypes` and
`CompilationDatabase` (but we do not have access to `Console` in
`KnownBuildRuleTypes`), there's no need to create a new one in
`AppleConfig`.  Passing one down lets us avoid having to schlep a
`Console` around.

I also added a new unit test to confirm the behavior in `AppleConfigTest`.
When I did this, I realized `AppleConfig` was using the non-testable API
of `ProcessExecutor` which takes an already-launched `Process` instance
instead of the testable API which takes a `ProcessExeuctorParams`, so I
fixed that.

Test Plan: Added new unit test. `ant java-test -Dtest.class=AppleConfigTest`.