[pathlib] Fix test to pass on Darwin
Summary:
The pathlib code assumes it'll get an `OSError` if there is an exception thrown,
but in reality, it doesn't appear to be that at least on macosx.  Additionally, they
don't re-raise the exception if it's for something that isn't what they were
looking for, which could result in failures being swallowed up.

Test Plan: CI on macosx