ndk_library: don't include absolute paths in `#makefile` rule key
Summary:
Previously, if a `ndk_library` had a `cxx_library` dep, we'd write
absolute paths into the generated makefile, which caused it to have
different contents and a different rule key per-user.  This diff
changes this to write a consistent makefile using relative paths,
and uses a makefile variable to forward the path to the project
directory to `ndk_build`.

Test Plan:
Added unittest to verify that makefiles generated in different working
directories generate identical rule keys.