Invalidate parser when android NDK path changes
Summary:
The code the generates the buckd daemon in `Main.java` has comments
which mention that the parser needs to be invalidated if the
`AndroidDirectoryResolver` changes, but this wasn't the case.  This
was causing issue when NDK libraries dependended on C/C++ libraries,
since the platform map that gets generated and passed into
`NdkLibraryDescription` objects uses the location of the NDK root.
This ends up getting cached in the cached `TargetNode` in the parser,
so changing the location of the NDK while buckd is still running
essentially had no effect.

This diff modifies the `equals` methods of the relevant components
so that a change in the NDK location triggers all the `TargetNode`
objects to get rebuilt with the updated descriptions.

Long term it'd be nice to never cache the old descriptions.  The
`TargetNode` objects can probably just hold references to the
descriptions they need and look these up from the repo factory when
they need to be used.  This would mean we wouldn't need to invalidate
the parser based on things that only affect descriptions.

Test Plan: Added a few tests.