Use AppleToolchainDiscovery in AppleSdkDiscovery
Summary:
This builds on my previous diff which added `AppleToolchainDiscovery`
to use it to discover all Xcode toolchains on the system and use them if
specified in the Apple SDK, falling back to the default if not found.

Since the `Toolchains` parameter in `SDKSettings.plist` is an array,
I made `AppleSdk.toolchains` into a `Set`, each of which is searched
for tools in order.

Test Plan: Unit tests updated.