Ignore SDKs with unrecognized platforms
Summary:
It's not safe to call `Enum.valueOf()` passing in an
arbitrary string from a file we read on disk.

Previously, this code would throw `IllegalArgumentException` for
any SDK with a platform that was unrecognized. With this diff, we'll
ignore any SDKs which have an unrecognized platform.

Test Plan: Unit test updated.