NdkCxxPlatform: support non-system C/C++ runtimes
Summary:
The default Android NDK C/C++ runtime is extremely minimal, so it's
common for users to choose a more full-featured one
(http://www.kandroid.org/ndk/docs/CPLUSPLUS-SUPPORT.html).  This
diff adds an option to the `NdkCxxPlatform` class to support choosing
an alternate runtime, which is used to select the GNU STL runtime
for the preset platforms in `KnownBuildRuleTypes`.

Test Plan:
Added a test that builds a C++ source, which uses standard library
components, and verifies it links.