Fix flaky DownloaderIntegrationTest
Summary:
Previously, `DownloaderIntegrationTest` would randomly
fail if its pseudo-random port number generator happened to choose
a port that was in use.

With this diff, we now let the system choose an unused port
by calling `AbstractNetworkConnector.setPort(0)`, and fetch
the local port number on which the server is listening after
starting the server.

Test Plan: ant java-test -Dtest.class=DownloaderIntegrationTest