Refactor TargetsCommand
Summary:
Previously, `TargetsCommand` used a predicate with a mutable state that had to
be applied on the graph nodes in the right order (dependencies before
dependents). This revision changes the implementation to first use a simpler
predicate to find direct owners of the referenced files, then use breadth
first search to find all nodes that depend on the direct owners and finally
filters the returned target according to the command line options.

Another benefit this revision brings is that we now will be able to fully
customize whether we just want to find all dependents of the direct owner or
whether we want to follow other graph edges as well.

This revision should not change the behavior of the `buck targets` command in
any way.

Test Plan: Unit and integration tests.