[ios] collect frameworks declared in test targets when generating combined tests
Summary:
Previously, the frameworks in the test libraries that are being combined were
not being listed in the frameworks section. This is often innocuous since the
library under test usually declares the required frameworks, but breaks in the
cases where they do not.

Test Plan: Modified test case to check for these kind of frameworks.