Remove project-level granularity when generating a combined project.
Summary:
Before this commit, when a workspace was generated and it depended on a target
that was a part of a project, all targets of that project were included in the
generated workspace. This diff changes that behavior for combined project. Since
workspaces generated with combined projects do not share their project files, we
can drop any extra targets that we don't depend on.

Test Plan: Generate a workspace with a combined project for an app and compile it.