Make native exopackage work on Android Lollipop.
Summary:
There are two issues with the symlink approach on Lollipop:
1. "ln -s" returns a "Permission denied" error.
2. Loading ".so" file from any location not under the app's data directory is not allowed.

To avoid these, this diff adds a new code path specifically for Lollipop.
- does not create any symlinks. Instead, the metadata.txt contains the mapping
from "library name" to the "native-<>.so" file in the /local/tmp/exopackage
directory.
- The exopackage loader copies ".so" files to an app-private directory called
"exo-libs" and loads libraries from that directory.
- The exopackage loader clears the "exo-libs" directory on start-up.

Test Plan:
```
$ buck build src/com/facebook/buck/android/support:buck-android-support
$ ./scripts/exopackage_test_single.sh
```

Tested on a Nexus 6 running Android 5.0.1, Galaxy Nexus running 4.2.2 and
Moto X running 4.4.4.