New class JsonObjectHashing
Summary:
This is split out from an earlier Parser diff I sent out,
and includes unit tests which were requested.

Note that this depends on and mimics the behavior of `RawParser` where
it changes many (but not all) numeric types to `Long`. (JSON numbers
are hard, m'kay?)

Test Plan: Unit tests included.