Limit generated clang header maps to 0.75 load factor
Summary:
Previously we don't resize the header map hash table till it's full. In the
event that we have exactly power of two number of entries, the entire table
would be full and on lookup miss, clang will loop forever looking for a null
entry to terminate the search. Not to mention that lookup performance suffers
as load factors increase.

This diff limits load factor to 0.75, which coincidentally seems to be what
xcode uses when generating its header maps.

Test Plan:
updated tests.
Ran on a buck target that has exactly 512 headers, and see that it generated a
header map with 1024 buckets instead of 512.