Implement --detect-test-changes for 'buck targets'
Summary:
Implement a new option (--detect-test-changes) for the `buck targets` command
that modifies the behavior of --referenced-file. If the new option is specified,
Buck will pretend that all targets depend on their tests when it determines
which targets reference the given files. This can be used to let us know if
tests of an app have been modified.

Test Plan:
  * Updated unit tests.
  * Run the command with the new option in a real codebase to confirm that it
    works correctly.