Shorten buck command lines
Summary:
Pass the bootstrap and regular class paths in the environment instead of on the command line. This way, the remainder of the command line fits under the Linux 4k /proc/pid/cmdline limit, making it possible to actually see what a given process is doing.

We filter our classpath environment variables out of the environment we use for subprocesses. (Java, amazingly, has `getenv`, but not `setenv` or `unsetenv`.)

Test Plan: buck build buck