Add a "locales" option to android_binary rule.
Summary:
This diff adds a new argument to the android_binary() rule, namely "locales",
which is a list of valid android locale codes, such as "en", "en_US", "es",
"es_ES" etc.

If the list of locales passed is empty, the behaviour does not change at all.
Otherwise, the filter resources step completely filters out "values-*" directories
for locales not in the list of locales passed.

Ideally, I'd like to pass this list of locales to the mini-aapt in order to
reduce the amount of work it does, but that means traversing down the transitive
dependencies of the android binary rule, and generating new android_resource rules
with a "locales" flavor. Perhaps some other day ...

Motivation behind the diff:
Many times developers remove a string from an english locale strings.xml file,
but forget to update strings.xml files for the rest of the locales in the app.
Now suppose there's a reference to the deleted string in some java code or a
layout xml file; the app still builds because aapt can find the deleted string
in some other locale file. However, when they eventually update the string files
or when they try to build with the "strings-as-assets" option, just the default
resources (values directory) are used to resolve resource references, and this
is when the build throws a "missing resource" error. This diff makes it so that
developers can define an english only build target for their debug build and
fix their mistakes earlier.

There's an added benefit here - aapt needs to only process the english resources
- thereby speeding up build times slightly.

Test Plan:
FilterResourcesStepTest.
AndroidResourceFilterIntegrationTest.

Build an app with `locales = ['en']`, switch locate to dutch and notice that
the app continues to use english strings.