Pull EnhancementResult out of AndroidBinaryGraphEnhancer.
Summary:
Immutables doesn't support inner classes / interfaces properly: the generated
javadoc references them by name only, and the generated immutable class is an
outer class, not an inner one. It is unclear at the moment whether those are
bugs or if it's not supposed to work at all.
For now we can work around this by pulling the inner interface out.

Test Plan: CI