Remove Flavor.DEFAULT.
Summary:
Flavor.DEFAULT is a leftover from when build targets had only one
flavour. The code now already treats an empty set of flavours as
equivalent to a set containing only the default flavour in some
places. This extends this equivalence to all the code base and
removes Flavor.DEFAULT.

Test Plan: CI