Extract logic for preparing a build report from the BuildCommand.
Summary:
By so doing, we can remove a static method from `BuildCommand` and make
it generally available as part of the `Build` itself. This decouple the
`TestCommand` from the `BuildCommand`, and makes it possible to run any
build outside of the `BuildCommand`, yet still get consistent output.

Test Plan: buck test --all