Do not throw spurious warnings when generating R.java for a locale specific APK.
Summary:
I added a new argument to the android_binary rule in a previous diff to filter
out values-* directories for specific locales. This is a follow up diff to get
rid of some spurious warnings.

While generating the uber R.java file, we gather resource symbols defined by
all the transitive android_resource dependencies of the APK, and re-assign the
ids using the R.txt generated by aapt. We also throw a warning if we do not
find a symbol in the uber R.txt generated by aapt. However, for locale filtered
builds, since mini-aapt processes _all_ the values-* directories, it is bound
to return symbols that aapt does not. Throwing a warning in this case is not
really useful, so this diff disables it for such builds.

Test Plan: build Facebook android apps (english only).