2x speedup for 'ant generate-immutables' by only running on subset of sources
Summary:
When running the Java annotation processor to generate Immutables
classes on our codebase, we were previously running it on every file in the
codebase.

Turns out this is twice as slow as just grep'ing the codebase for
'org.immutables' and then running the annotation processor on those
files.

Test Plan:
Before:

  % time ant generate-immutables
  Buildfile: /Users/bgertzfield/devtools/buck/build.xml

  generate-immutables:
      [javac] Compiling 856 source files

  BUILD SUCCESSFUL
  Total time: 5 seconds
  ant generate-immutables  11.80s user 0.55s system 233% cpu 5.295 total

After:

  % time ant generate-immutables
  Buildfile: /Users/bgertzfield/devtools/buck/build.xml

  generate-immutables:
      [javac] Compiling 15 source files

  BUILD SUCCESSFUL
  Total time: 2 seconds
  ant generate-immutables  5.34s user 0.28s system 208% cpu 2.701 total