100% speedup for 'ant compile-bootstrapper' when no sources have changed
Summary:
When running `ant compile-bootstrapper`, I noticed it always
compiled all source files, even if they hadn't changed. This took about
1.5 seconds on every incremental compile.

Running `ant -d compile-bootstrapper`, I saw this curious output:

  [javac] ClassLoaderBootstrapper.java added as ClassLoaderBootstrapper.class doesn't exist.
  [javac] ConsoleHandler.java added as ConsoleHandler.class doesn't exist.
  [javac] LogConfig.java added as LogConfig.class doesn't exist.
  [javac] LogFormatter.java added as LogFormatter.class doesn't exist.
  [javac] Compiling 4 source files to /Users/bgertzfield/devtools/buck/build/bootstrapper/classes

I spent a while debugging, and eventually figured out this was because
we were passing to Ant's `javac` task a `srcdir` argument pointing to
the specific subdirectory containing just the bootstrapper sources.

Unfortunately, this bypassed ant's built-in check to see if compiled
`.class` files existed in the proper subdirectories, since it didn't
grok the package layout.

To fix this, this diff ensures that we run the `javac` with `srcdir`
set to the default (the top-level `src` directory) and instead rely
on `includes` to pull in the `srcdir`-relative bootstrapper sources
and not compile anything else.

Test Plan:
  % repeat 2 time ant compile-bootstrapper
  Buildfile: /Users/bgertzfield/devtools/buck/build.xml

  checkversion:

  compile-bootstrapper:
      [mkdir] Created dir: /Users/bgertzfield/devtools/buck/build/bootstrapper/classes
      [javac] Compiling 4 source files to /Users/bgertzfield/devtools/buck/build/bootstrapper/classes
      [javac] [compiled 177 lines in 536 ms: 330.2 lines/s]
      [javac] [4 .class files generated]

  BUILD SUCCESSFUL
  Total time: 1 second
  ant compile-bootstrapper  2.76s user 0.15s system 186% cpu 1.563 total
  Buildfile: /Users/bgertzfield/devtools/buck/build.xml

  ===

  checkversion:

  compile-bootstrapper:

  BUILD SUCCESSFUL
  Total time: 0 seconds
  ant compile-bootstrapper  1.12s user 0.08s system 170% cpu 0.699 total