[Fix] fix ManifestMerger fail to create XmlFile with #
Summary:
There is a bug / feature in `XmlUtils.printXmlFile` which will truncate a
file name if it contains `#`. Its kind of valid since `#` is not allowed
in XML file name, but we are using `#` as delimiter in buildTarget if it's
Flavored.
For fixing that, instead of generating the file directly, generate the
XmlString first, and then write the data into the file.

Test Plan:
1. build android_manifest under a folder with `#`, output doesn't get truncated
2. arc unit