Add jackson-datatype-jdk7 to repo so we can serialize Path objects
Summary:
There's been a longstanding TODO in AuditInputCommand to make
Jackson serialize `Path` objects.

Simon filed a task way back in December of 2013 to make this happen:

https://github.com/FasterXML/jackson-databind/issues/373

but we never followed up to use the new library in Buck which the
Jackson author added for us.

This adds the dependent library `jackson-datatype-jdk7` to the repo so
we can make this happen.

Test Plan: Used from subsequent diff.