CxxPlatform: Annotate interface to make it Immutable
Summary:
This annotates the `CxxPlatform` interface so we auto-generate
an `Immutable` implementation and builder.

This does *not* change the existing implementations (`DefaultCxxPlatform`,
`NdkCxxPlatform`, `AppleCxxPlatform`). That will happen next; in the short
term, we have both `Immutable` and manual implementations of `CxxPlatform`.

I generalized the return types so the Immutables code-generator
understood the `List` and `Multimap`.

I had to change a few downstream APIs in `OCamlBuckConfig` since they
depended on the old interface.

Test Plan:
  % ant java-test -Dtest.class=OCamlIntegrationTest
  % ant java-test -Dtest.class=DefaultCxxPlatformTest