Remove catch all case in the rule key builder.
Summary:
The rule key builder has a case in setReflectively that catches any
value that isn't handled explicitly. This makes it hard to tell
whether a certain type of value is handled correctly or not. This
remove that and introduces explicit handling for build targets, enums
and strings.

Test Plan: CI, build app, verify all rules are cache hits