Allow passing 'is executable' checker to MoreFiles.searchPathsForExecutable
Summary:
We currently have no way to mock out the filesystem behavior of
`MoreFiles.searchPathsForExecutable()`.

We can't use `ProjectFilesystem` for this, since we're searching for files
outside the project root.

So, this allows test callers to provide a `Function<Path, Boolean>` to check
if a path exists and is executable.

Test Plan: Unit tests included.