Fix construction of TargetSources
Summary:
Previous diff broke project generation for targets that had sources with an
extension that was neither a known header file extension nor a known
implementation file extension. This fixes the issue by explicitly returning a
set of all source files regardless of whether we recognize their extension
or not.

Test Plan: CI