Use either the hash of the external javac or its version information.
Summary:
Java is Java is Java. It is assumed that if two versions of what are meant
to be the same version of a javac (such as one running on Windows, and one
on Linux) they'll emit the same bytecode. For this reason, if version info
is given on stderr use that in preference to hashing the compiler. Despite
this, we should fall back to hashing the compiler when version information
is missing.

Test Plan: buck test --all