Ensure bridge methods are included in stub jars
Summary:
If you're compiling against stub jars generated by the `ClassMirror`
it's necessary for bridge methods to be included.

It's possible that the signatures (and therefore the rule keys) from
classes that were meant to have bridge methods are wrongly generated
which obviously causes problems.

Test Plan: buck test --all