Gracefully handle `Thread.interrupt()` during parsing
Summary:
Previously, when we received an `InterruptedException` while parsing a
BUCK file, this would be thrown as a flavor of `IOException` which we
converted into a parse exception.  This diff catches the specific types
of `IOException` throw when an I/O operation was interrupted, and propates
this up as an `InterruptedException` so that we handle this as correctly.

Test Plan:
We changed the order of some operations for `buck test` which exposed
this issue in the `DaemonIntegrationTest.whenClientTimeoutDetectedThenTestIsInterrupted`
test.  This diff fixes the underlying issue and the test.