Rename AppleSource to SourceWithFlags
Summary:
Rename `AppleSource` and related classes to `SourceWithFlags`, because no part
of it is specific to Apple targets and it can be used in other contexts as well.
For example, I am going to use the renamed class for sources in
`CxxConstructorArg`, so we can specify per-file flags in cxx targets.

Test Plan: `buck test`