Adds an `audit dependencies` command to the CLI
Summary:
This command takes a list of targets as input and prints all of the
targets that those targets depend on. It only prints the immediate
dependencies by default, but using the `--transitive` flag will
cause it to include transitive dependencies as well. This is mainly
useful for automation and small scripts.

This also fixes the incorrect existing documentation. It specified
that `input` and `classpath` were flags, when really they are full
on commands. Instead of adding fuel to the fire and adding incorrect
documentation for this new command, I instead just updated the audit
documentation to reflect the way the command is currently used.

Test Plan:
* Unit Tests.
* Run soy2html, then use SimpleHTTPServer to view the webpage. Confirm that everything looks decent.