TestCommand: don't build filtered tests by default
Summary:
We changed the default for `buck test --all` so that filtered
tests were built be defult, even though we didn't run them during the
test.  This result in a *lot* of `buck test --all` commands suddenly
taking significantly longer.

This diff changes the default back to not building filtered tests by
default and adds a `.buckconfig` setting to enable this behavior.
This default generally seems nicer as well, since it'll do the minimum
work required for the test command.

Test Plan:
- adjusted test