Introduce PMD.LooseCoupling check.
Summary:
In practice, this encourages us to refer to collections via their
interface (`Map`) rather than their underlying collection type (`HashMap`).
A number of source files (12) had to be updated to satisfy this new lint check.

I came across this while checking to see if it would be reasonable to
introduce static checking for the Law of Demeter. It turns out that PMD
has a built-in static check for this, but flags 21210 rule violations in
our code, so I don't see us cleaning those up. (Also, my gut feeling has
always been that the Law of Demeter is too strict, in practice. This
confirms my suspicions.)

Test Plan: Sandcastle builds.