Improve debugging of unresponsive threads
Summary:
Occasionally buck threads used for building can become unresponsive.
We eventually time these out (after a ridiculous 10m), but even then
we don't get any insight into what went wrong.

This diff moves the thread pool management to the `Command` classes
and adds a new `CommandThreadManager` which encapsulates the thread
pools command use.  In particular, it'll keep track of the `ThreadGroup`
used to create all of it's threads and use that to print a more
informative error message which includes the stack trace of all blocked
threads.

Test Plan:
- added unittest for `CommandThreadManager`
- adjusted unittest for `DefaultStepRunnerTest`
- explicitly added a thread to the build process which just spun and
  verified it timed out with a useful error message during a build