Revert "Normalize SourcePath handling in RuleKeys."
Summary:
This breaks caching for `BuildRule`s which get the same `PathSourcePath`
from different path names.  Reverting since I'm not sure of a great
way forward.  I kind of like the idea of dropping the path name altogether
and requiring `BuildRule` objects to explicitly add them to the rule
key if they're important, but this will require some chages (to at least
`AndroidResource`).

Test Plan:
Build a cxx_binary rule which uses an in-repo preprocessor tool.  This
tool gets the full path when added to the rule key.