Properly handle explicit exclusions on Windows
Summary:
This refactors our glob tests to test the Windows and Posix code
paths on both platforms.  It also adds a test to catch the case
where we have a file separator in our excludes with an explicit
file path that is not evaluated as a glob pattern, which fails
on Windows.

Addresses part of https://github.com/facebook/buck/issues/248

Test Plan: `ant buck-py-test`