Refactor ProjectGraphParser into its own file
Summary:
This is a follow-up to my previous diff
`Refactor ProjectCommand logic to create TargetGraph parser so it can be reused`.

It splits out the interface `ProjectGraphParser` into its own file with
a utility class `ProjectGraphParsers` to create an instances of that interface.

The next step will be to use this logic from `TargetsCommand`.

Test Plan: `ant java-test -Dtest.class=ProjectCommand*Test`