Add ability to use a .src.zip as a python_library source.
Summary:
The intention is to allow genrules to generate > 1 file worth of code (or some directory structure) that can be used in a python_library. For example, you could write a small wrapper around genrule to take a PyPI package name & version, download it & repackage it to include certain directories as a python_library (similar to, but slightly more involved than maven_jar in bucklets).

Closes https://github.com/facebook/buck/pull/264

Test Plan: `buck test` with a new integration test.