HttpArtifactCache: use a cheaper hash for checksumming
Summary:
Previously, we'd pass in a `FileHashCache` object to provide the
hash code for the file we were fetching/storing.  However, by default
this does a sha1, which is really expensive and overkill when just
trying to detect data corruption.

Also since we're pushing artifacts to/from the cache that never were
hashed by build rules, using a `FileHashCache` never helps us, as we'll
never get a hit and avoid the operation.

This diff switches to a significantly cheaper hash code and also moves
the hashing calculation to be done *before* we initiate the store to
the cache, so that we hold the connection open for less time.

Test Plan:
Ran a build locally in HTTP readwrite mode.  Also, used a variety of different
hash codes as provided by `Hashing` and `crc32` was by farther fastest.